Final commit.

+++ backup_and_transfer.h
@@ -44,12 +44,20 @@ void backup_and_transfer(char *folder, char *bak_loc, char *trans_loc) {
     snprintf(command, sizeof(command), "tar Pczf %s/%s %s", bak_loc, filename, folder);
     syslog(LOG_DEBUG, "Backup command: %s", command);
     syslog(LOG_INFO, "Creating backup: %s", filename);
-    system(command);
+    if (system(command) != 0) {
+        syslog(LOG_ERR, "Failed to run command.");
+        send_msg("FAILURE: Failed to create backup.");
+        exit(EXIT_FAILURE);
+    }
 
     syslog(LOG_INFO, "Beginning transfer to: %s", trans_loc);
     snprintf(command, sizeof(command), "rsync -cr %s/ %s/ --delete-after", folder, trans_loc);
     syslog(LOG_DEBUG, "Transfer command: %s", command);
-    system(command);
+    if (system(command) != 0) {
+        syslog(LOG_ERR, "Failed to run command.");
+        send_msg("FAILURE: Failed to transfer.");
+        exit(EXIT_FAILURE);
+    }
 
     syslog(LOG_INFO, "Unlocking folder.");
     change_perm(folder, "0777");
